Добавлена привязка контроллера с входом и исправлен баг изменением имени контроллера

+++ app/src/main/java/ru/cityron/presentation/screens/controller/ControllerSettingsViewModel.kt
@@ -3,12 +3,13 @@ package ru.cityron.presentation.screens.controller
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import ru.cityron.domain.repository.CurrentRepository
-import ru.cityron.domain.usecase.controller.DeleteControllerUseCase
 import ru.cityron.domain.usecase.all.settings.eth.GetM3SettingsEthUseCase
 import ru.cityron.domain.usecase.all.settings.http.GetM3SettingsHttpUseCase
 import ru.cityron.domain.usecase.all.settings.metric.GetM3SettingsMetricUseCase
 import ru.cityron.domain.usecase.all.settings.time.GetM3SettingsTimeUseCase
 import ru.cityron.domain.usecase.all.state.GetM3StateUseCase
+import ru.cityron.domain.usecase.bind.BindDeleteUseCase
+import ru.cityron.domain.usecase.controller.DeleteControllerUseCase
 import ru.cityron.presentation.mvi.BaseSharedViewModel
 import javax.inject.Inject
 
@@ -21,6 +22,7 @@ class ControllerSettingsViewModel @Inject constructor(
     private val getM3SettingsHttpUseCase: GetM3SettingsHttpUseCase,
     private val currentRepository: CurrentRepository,
     private val deleteControllerUseCase: DeleteControllerUseCase,
+    private val bindDeleteUseCase: BindDeleteUseCase
 ) : BaseSharedViewModel<ControllerSettingsViewState, Any, ControllerSettingsViewIntent>(
     initialState = ControllerSettingsViewState()
 ) {
@@ -85,8 +87,10 @@ class ControllerSettingsViewModel @Inject constructor(
     private fun onDeleteClick() {
         withViewModelScope {
             currentRepository.current.value?.let {
-                deleteControllerUseCase(it.id)
-                currentRepository.setCurrentController(null)
+                if (bindDeleteUseCase(it.num)) {
+                    deleteControllerUseCase(it.id)
+                    currentRepository.setCurrentController(null)
+                }
             }
         }
     }
